mspi: cadence: allow AHB trigger offset configuration
For indirect read/write operations, data is transferred using a FIFO in the
flash region by the controller. Currently this is configured to be at
offset 0 i.e, at the start of the flash region. The user might want to
change the offset to something else for a variety of reasons, and hence,
allow configuring this region by modifying the following DT properties.

- 'reg' property at index 1 now contains the base address for the flash
  region.
- 'indirect-trigger-offset' property is used to set the offset for the FIFO
  making the absolute address for the fifo as the sum of this value and the
  aforementioned address.

The reason we need both these properties is that, we need the offset for
configuring the controller register and the absolute address (sum) for host
access.

Signed-off-by: Amneesh Singh <amneesh@ti.com>